requirements.txt
pytest~=8.3.2
selenium~=4.23.1
allure-python-commons~=2.13.5
PyYAML~=6.0.2